requirements.txt
Django==1.11.5
django-ckeditor==5.3.0
factory-boy==2.9.2
django-extensions==1.9.0
django-admin-sortable==2.1.1